Changed action.parseFloat() to action.parseInt()

+++ ArduinoCode/monochromatorControl2/monochromatorControl2.ino
@@ -141,7 +141,7 @@ void loop() {
   while (!Serial.available());
   //delay(300);
   if(Serial.available()){
-    action = Serial.parseFloat();
+    action = Serial.parseInt();
   }
